chore: move attributes to a new file
